requirements.txt
Django==2.0.8
django-tenants==2.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python-decouple==3.1
pytz==2018.7
